% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/dataEdit.R
\docType{class}
\name{dataEdit-class}
\alias{dataEdit}
\alias{dataEdit-class}
\title{Reference Class for  Editing Data Table Objects}
\description{
A subclass of \code{\link{dataTable-class}} objects, in which elements in a column can be edited by supplying
new values for a range of rows, and the edits can be undone.
}
\section{Fields}{

\describe{
\item{\code{edits}}{A list that keeps account of the edits performed.}

\item{\code{description}}{A character string with a description of the edit project.}
}}
\section{Methods}{

\describe{
\item{\code{edit(i, var, value)}}{Replaces the range i in the variable named var in
the object by value.}

\item{\code{initialize(..., data = data.frame())}}{This method takes \code{data=} as a named argument, but coerces it to be an environment,
instead of requiring that it already be one.}

\item{\code{undo()}}{Undoes the last edit() operation
and updates the edits field accordingly.}
}}

