% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataTable.R
\docType{class}
\name{dataTable-class}
\alias{dataTable}
\alias{dataTable-class}
\title{Reference Class for Data Frame-like Objects}
\description{
Objects from this class behave rather like \code{\link{data.frame}} objects, but are formal reference classes
so that computations with them use reference semantics
}
\section{Fields}{

\describe{
\item{\code{data}}{An environment whose objects correspond to the "columns" of a data frame or matrix.}

\item{\code{row.names}}{The names for the "rows", treated as those in a data frame.}
}}
\section{Methods}{

\describe{
\item{\code{initialize(..., data = data.frame())}}{This method takes \code{data=} as a named argument, but coerces it to be an environment,
instead of requiring that it already be one.}
}}

