% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/datedClasses.R
\docType{class}
\name{datedText-class}
\alias{datedList}
\alias{datedList-class}
\alias{datedTable}
\alias{datedTable-class}
\alias{datedText}
\alias{datedText-class}
\alias{datedVector}
\alias{datedVector-class}
\title{Classes Adding DateTime slot/field to Basic Classes}
\description{
These classes all add a simple \code{date} slot or field to data
of a basic type.  As described in Chapter 10, class \code{"datedTable"}
is a reference class, with a field for the environment data.
}
\details{
All the classes have corresponding generating functions of the same name
that will take \code{date=} arguments for that slot/field, unnamed arguments
for the data part, except for  \code{datedTable} which needs \code{table=}
for the environment to use.
}
\section{Slots}{

\describe{
\item{\code{date}}{The date information, using class \code{"DateTime"} to include
any of the classes for representing such information.}
}}

