% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/montyHall.R
\name{montyHall}
\alias{montyHall}
\title{Simulating the "Monty Hall" Game}
\usage{
montyHall(N)
}
\arguments{
\item{N}{sample size; i.e., number of times to run the simulation of the game.}
}
\value{
A data frame with columns for the first guess, the opened
door/box, the second guess and the location of the prize.
}
\description{
In the game, a prize is located behind one of 3 doors (or in one of 3 boxes). The
player makes a guess, after which the game owner opens one door/box that is neither
the winner nor the guess.  The player can then make a second guess.  This function
runs the game \code{N} times.
}
\details{
The assumption is that both the prize location and the first guess are sampled
uniformly.  Also, if the owner has a choice, the open is also sampled.

The point is to show that the second guess has twice the winning chance of the
first.  Sometimes framed as Bayesian inference (Bain, 2016) but in fact
obvious if the event space is partitioned relevantly (as the example
illustrates).
}
\examples{
set.seed(378)
N <- 10000
mh <- montyHall(N)
x <- mh[mh$guess1 != mh$prize,] # guess1 was wrong
all(x$guess2 == x$prize) # => guess2 was right
sum(mh$guess1 == mh$prize)/N ## probability 1/3
sum(mh$guess2 == mh$prize)/N ## probability 2/3
}
\author{
John M. Chambers
}
\references{
Chambers, John M. (2016)
\emph{Extending R},
Chapman & Hall/CRC.

Bain, Robert. (2016)
\dQuote{Are our brains Bayesian?}
\emph{Significance}, vol. 13, issue 4, pp 14-19.
}
\keyword{Programming}

