\name{XRexamples-package}
\alias{XRexamples-package}
\alias{XRexamples}
\docType{package}
\title{
\packageTitle{XRexamples}
}
\description{
\packageDescription{XRexamples}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{XRexamples}
\packageIndices{XRexamples}

}
\author{
\packageAuthor{XRexamples}

Maintainer: \packageMaintainer{XRexamples}
}
\keyword{ package }
\keyword{ programming }
